Re-structure the run_command method
First, I like passing in %params instead of all the extra variables. SSH
is a little different so I kept it as-is.

Second, I felt that we should just do the eval stuff inside of the
run_command method. This way, we can just pass back a `error` and check
for that in the results. Also, we don't really have to worry about
returning the output in the run_command method.

Third, I think that we shouldn't require the STDIN. We should just pass
in an empty-string if it is not passed in.